[#6] refactor: 리다이렉트 uri 상수로 분리
