+++ src/lib/data.ts
@@ -34,17 +34,3 @@ export const SORT_OPTIONS = [
   { type: 'createAt', name: '최신순' },
   { type: 'title', name: '제목순' },
 ] as const;
-
-const KAKAO_LOGIN_KEY = import.meta.env.VITE_KAKAO_LOGIN_API_KEY;
-export const KAKAO_REDIRECT_URI = import.meta.env.VITE_KAKAO_LOGIN_REDIRECT;
-
-export const KAKAO_AUTH_URL = `https://kauth.kakao.com/oauth/authorize?client_id=${KAKAO_LOGIN_KEY}&redirect_uri=${KAKAO_REDIRECT_URI}&response_type=code`;
-
-const GOOGLE_LOGIN_KEY = import.meta.env.VITE_GOOGLE_LOGIN_API_KEY;
-export const GOOGLE_REDIRECT_URI = import.meta.env.VITE_GOOGLE_LOGIN_REDIRECT;
-
-export const GOOGLE_AUTH_URL = `https://accounts.google.com/o/oauth2/v2/auth?
-client_id=${GOOGLE_LOGIN_KEY}
-&redirect_uri=${GOOGLE_REDIRECT_URI}
-&response_type=code
-&scope=https://www.googleapis.com/auth/indexing`;
